scripts/requirements.txt
requests
Pyyaml
pytablewriter
jinja2
yarl
click
attrs


# tooling
pylint
